Add 1D advection-diffusion of passive tracer
Have added the functions to do this but the way the multiple dispatch is used is not so good. Need to write some tests to make sure that everything is working correctly then check with Navid if there is a better to way to set this up. I think doing it this way generalises very easy to the 3D domain.